requirements.txt
spacy
requests
python-dotenv
nltk
geocoder
